+++ gatsby-site/src/components/blog-menu/BlogMenu.js
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import blogList from '../blog-list/BlogList'
+import styles from "./BlogMenu.css"
+
+export default function BlogMenu() {
+
+  return (
+      <div className="blog-menu">
+          {blogList}
+      </div>
+    );
+
+}
